Added clipping of the tangle at runtinme, rather than just in the initialisation.
This is important for high velocities and time steps which are probably a bit too large. In this case
more than one vortex node can move past a wall in a single time step. The reconnection algorithm can
handle one point moving out, however, when this happens for more points a loop
is created on the opposite side of the wall. These non-physical loops are trimmed now and a warning
message is printed. This problem can be avoided by smaller time step, however, this might end up being
prohibitively small.