New initial conditions and configuration options.
Hyperfriction and maximum number of steps in the simulation is now settable through the config file.

New initial conditions -- vortex dipole and vortex quadrupole in slab geometry.

A minor tweak to add_point() in remeshing to not try to extrapolate point placement with very strong curvature.
It helped a bit, but there is still spurious loop emission when a wall reconnection occurs.